Move PolygonExtensions alongside their interfaces.
